Function References

Surrogate

NMRSignalSimulator.CLSurrogateConfig — Type

struct CLSurrogateConfig{T} Δr::T = convert(T, 1.0) Δκλ::T = convert(T, 0.05) Δcsmaxscalar::T = convert(T, 0.2) κλlb::T = convert(T, 0.5) κλ_ub::T = convert(T, 2.5)

default_ppm_padding::T = convert(T , 0.5)

end

default_ppm_padding –

κ_λ_lb, κ_λ_ub – the default lower and upper bounds, respectively, for the κ_λ input of the surrogate.

Δcs_max_scalar, default_ppm_padding – specifies the surrogate's default frequency operating range, [umin - deltau, umax + deltau], where deltau is Δcsmax_scalar expressed in Hz.

umin, umax are in Hz, Δcsmaxscalar is in ppm. u_min is set to the smallest resonance frequency from As, and u_max is set to the largest resonance frequency from As. Δcs_max_scalar is in units of ppm. It is the border that is added to umin and umax (once they are converted to ppm) to get the final frequency interval for which the surrogate of a spin system is fitted to. The extrapolation of the surrogate for frequency queries outisde this interval is set to return zero, so the surrogate is the zero signal outside this interval for that spin system.

Δr, Δκ_λ – the sampling increment for the frequency input r and T2 multiplier input κ_λ for generating samples to fit the surrogate. Smaller means the surrogate is more accurate, but slower to construct the surrogate.

NMRSignalSimulator.fitclproxies — Function

fitclproxies(
    As::Vector{HAM.SHType{T}},
    λ0::T,
    config::CLSurrogateConfig{T};
    names::Vector{String} = Vector{String}(undef, 0),
    ) where T <: AbstractFloat

Create surrogate for NMR spectrum, given the simulated resonance comopnent results from NMRHamiltonian.

Inputs:

As – a 1-D array of compound resonance simulations. See NMRHamiltonian.simulate.

λ0 – the estimated T2* decay for the 0 ppm resonance component.

config – the configuration for building the surrogate. See CLSurrogateConfig.

NMRSignalSimulator.evalclmixture — Function

function evalclmixture(
    u_rad,
    As::Vector{HAM.SHType{T}},
    Bs::Vector{MoleculeType{T,SST,CLOperationRange{T}}};
    w::Vector{T} = ones(T, length(As)),
)::Complex{T} where {T <: Real, SST}

Evaluates the preliminary model at radial frequency (in radians) u_rad. Inputs:

Bs is the surrogate model.

As is the resonance component data structure.

w is relative concentration of each compound entry.

NMRSignalSimulator.evalclproxymixture — Function

function evalclproxymixture(
    u_rad,
    As::Vector{HAM.SHType{T}},
    Bs::Vector{MoleculeType{T,SST,CLOperationRange{T}}};
    w::Vector{T} = ones(T, length(As)),
)::Complex{T} where {T <: Real,SST}

Evaluates the surrogate model at radial frequency (in radians) u_rad. Inputs:

Bs is the surrogate model.

As is the resonance component data structure.

w is relative concentration of each compound entry.

Constructors

NMRSignalSimulator.getSpinSysParamsdatatype — Function

function getSpinSysParamsdatatype(
    ::Type{CoherenceShift{T}}
    )::DataType where T <: AbstractFloat

Convinence constructor for SpinSysParams{CoherenceShift{T}, CoherencePhase{T}, SharedT2{T}}

function getSpinSysParamsdatatype(
    ::Type{SharedShift{T}}
    )::DataType where T <: AbstractFloat

Convinence constructor for SpinSysParams{SharedShift{T}, CoherencePhase{T}, SharedT2{T}}

NMRSignalSimulator.MixtureModelParameters — Type

function MixtureModelParameters(
    MSS::CLMixtureSpinSys{T,ST,PT,T2T};
    w = ones(T, getNentries(MS)),
) where {T,ST,PT,T2T}

Convinence constructor for MixtureModelParameters(). Does not create a copy of of the inputs.

Model updates

NMRSignalSimulator.importmodel! — Function

function importmodel!(p::MixtureModelParameters)::Nothing

Updates p.MSS with the contents of p.var_flat.

function importmodel!(p::MixtureModelParameters, x, inds::SubsetVarsIndices)::Nothing

Only update the entries as specified by inds.

function importmodel!(p::MixtureModelParameters, x, ::AllVars)::Nothing

Calls importmodel!(p, x).

function importmodel!(p::MixtureModelParameters, x)::Nothing

Update the model p using the contents of x.

NMRSignalSimulator.importmodelreset! — Function

function importmodelreset!(p::MixtureModelParameters, x, inds::SubsetVarsIndices)::Nothing

This is a version of importmodel!(p::MixtureModelParameters, x, inds::SubsetVarsIndices)::Nothing that resets the other variables not mentioned in inds.

First, resets the model to default parameter values; 0 for shift and phase parameters, 1 for T2 multiplier parameters. Then update the model parameters as specified by inds with the entries in x.

NMRSignalSimulator.exportmodel! — Function

function exportmodel!(p::MixtureModelParameters)::Nothing

Flattens the model parameters into p.var_flat.

Subset vars

NMRSignalSimulator.VariableSetTrait — Type

abstract type VariableSetTrait end

NMRSignalSimulator.AllVars — Type

struct AllVars <: VariableSetTrait end

NMRSignalSimulator.SubsetVars — Type

struct SubsetVars <: VariableSetTrait
    indices::SubsetVarsIndices
    active_systems::Vector{Tuple{Int,Int}}

NMRSignalSimulator.SubsetVarsIndices — Type

struct SubsetVarsIndices
    shift::Vector{Int}
    phase::Vector{Int}
    T2::Vector{Int}

NMRSignalSimulator.exportindices — Function

function exportindices(A::SubsetVarsIndices)::Vector{Int}

Returns [A.shift; A.phase; A.T2]

function exportindices(A::SubsetVars)::Vector{Int} Returns index in A.

NMRSignalSimulator.setupsubsetvars — Function

function setupsubsetvars(
    indicies_input::Vector{Int},
    mapping::ParamsMapping;
)::SubsetVars

Convinence constructor for SubsetVars.

NMRSignalSimulator.getindices — Function

function getindices(A::AllVars)::AllVars returns A.

function getindices(A::SubsetVars)::SubsetVarsIndices returns A.indices.

NMRSignalSimulator.getactivesystems — Function

function getactivesystems(A::AllVars)::AllVars returns A.

function getactivesystems(A::SubsetVars)::Vector{Tuple{Int,Int}} returns A.active_systems.

Parameter range

NMRSignalSimulator.CLOperationRange — Type

struct CLOperationRange{T}
    u_min::T # in Hz.
    u_max::T # in Hz.
    
    # d_max = ppm2hzfunc.(Δcs_max) .- ppm2hzfunc(zero(T))
    d_max::Vector{T} # in Hz. ζ is in rad.

    κ_λ_lb::T
    κ_λ_ub::T
    Δr::T
    Δκ_λ::T
end

Operation range for the spline surrogate q(r,λ) of a resonance group. Bounds on r for groups in the i-th spin system:

r_min = 2*π*(u_min - d_max[i])
r_max = 2*π*(u_max + d_max[i])

Lower and upper bounds on λ are κ_λ_lb and κ_λ_ub, respectively. Able to serialize/deserialize.

NMRSignalSimulator.fetchbounds — Function

function fetchbounds(
    p::MixtureModelParameters,
    Bs::Vector{MoleculeType{T, SST, OT}};
    shift_proportion = 0.9, # between 0 to 1. Control the returned shift bounds as a proportion of the maximum allowed shift bounds used when the surrogates were created.
    phase_lb = convert(T, -π),
    phase_ub = convert(T, π),
)::Tuple{Vector{T},Vector{T}} where {T,SST,OT}

Returns a Vector for lower bound and Vector for upper bound for each parameter variable. The order (first elements to last elements) of the returned vectors are: Shift (ζ), phase (κβ), then T2 (κλ) parameters.

Input:

p – model parameters. Used to determine the size of the output.

Bs – Surrogate model. This function uses the CLOperationRange field from each element of Bs.

Optional inputs:

shift_proportion – between 0 to 1. Control the returned shift bounds as a proportion of the maximum allowed shift bounds used when the surrogates were created.

phase_lb – the fill value for the lower bound of phase parameter (κ_β).

phase_ub – the fill value for the lower bound of phase parameter (κ_β).

NMRSignalSimulator.MoleculeParamsMapping — Type

struct MoleculeParamsMapping
    st::Vector{Vector{Int}}
    fin::Vector{Vector{Int}}
end

NMRSignalSimulator.ParamsMapping — Type

struct ParamsMapping
    shift::MoleculeParamsMapping
    phase::MoleculeParamsMapping
    T2::MoleculeParamsMapping
end

NMRSignalSimulator.getshiftrange — Function

function getshiftrange(A::ParamsMapping)

Returns the range for the shift parameters (ζ).

NMRSignalSimulator.getphaserange — Function

function getphaserange(A::ParamsMapping)

Returns the range for the phase parameters (κ_β).

NMRSignalSimulator.getT2range — Function

function getT2range(A::ParamsMapping)

Returns the range for the T2 parameters (κ_λ).

Utilies

NMRSignalSimulator.getNentries — Function

function getNentries(MSS::CLMixtureSpinSys)::Int

Returns the number of molecule entries associated with MSS

function getNentries(model_params::MixtureModelParameters)::Int

Return the number of molecule entries associated with model_params

NMRSignalSimulator.exportphysicalparams — Function

exportphysicalparams(
    MSS::CLMixtureSpinSys,
    Phys::Vector{HAM.PhysicalParamsType{T}}, # template.
    fs::T,
    SW::T,
    ν_0ppm::T,
) where T

Outputs: outshift, outphase, out_T2.

The fields cs_sys and cs_singlets for out_shift, out_phase, and out_T2 does not contain chemical shift. It contains the shift difference (in ppm), phase variable κ_β (in radians), and the T2 variable ξ (a multiplier, dimensionless), respectively.

NMRSignalSimulator.createtablecolumns — Function

function createtablecolumns(Phys::Vector{NMRHamiltonian.PhysicalParamsType{T}})::Tuple{Vector{Vector{Int}},Vector{Tuple{Int,Int}},Vector{T}} where T

Extract information from Phys to produce three 1-D arrays that is useful for constructing tables for visualization purposes.

Outputs:

nucleus_ID_set – nucleus ID numbers, as recorded from Phys[n].H_IDs, i.e., the original not the re-labelled IDs.

location_set – pairs of the format (n,i), n is the compound entry index, i is the spin system index.

cs_set – chemical shift numbers for each nucleus, in units ppm.

NMRSignalSimulator.ζ2Δcs — Function

function ζ2Δcs(ζ::T, ν_0ppm::T, hz2ppmfunc)::T where T

Convert radial frequency to ppm. '''

# test.

a = Δcs2ζ(0.1, ppm2hzfunc) @show ζ2Δcs(a, ν_0ppm, hz2ppmfunc) '''

NMRSignalSimulator.Δcs2ζ — Function

function Δcs2ζ(Δcs::T, ppm2hzfunc)::T where T

Convert ppm to radial frequency. '''

# test.

a = Δcs2ζ(0.1, ppm2hzfunc) @show ζ2Δcs(a, ν_0ppm, hz2ppmfunc) '''

Serialize/Deserialize

NMRSignalSimulator.serializclproxies — Function

function serializclproxies(
    Bs::Vector{MoleculeType{T,SpinSysParams{ST,PT,T2T},OT}},
) where {T,ST,PT,T2T,OT}

Returns a dictionary containing the surrogate model.

NMRSignalSimulator.serializitpsamples — Function

function serializitpsamples(
    itp_samps_set::Vector{Vector{InterpolationSamples{T}}},
) where T

Returns a dictionary of the interpolation samples used to fit the surrogate.

NMRSignalSimulator.deserializclproxies — Function

function deserializclproxies(W)

W is of a data type that can be addressed via a key (of type Symbol) and returns a blue. Examples: type Dict{Symbol, Any} or JSON.Object.

NMRSignalSimulator.deserializitpsamples — Function

function deserializitpsamples(W)

W is of a data type that can be addressed via a key (of type Symbol). Examples: type Dict{Symbol, Any} or JSON.Object.

NMRSignalSimulator.recoverclproxies — Function

function recoverclproxies(
    itp_samps_set::Vector{Vector{InterpolationSamples{T}}},
    ss_params_set::Vector{SpinSysParams{ST,PT,T2T}},
    op_range_set::Vector{CLOperationRange{T}},
    As::Vector{HAM.SHType{T}},
    λ0::T,
    )::Tuple{Vector{MoleculeType{T,SpinSysParams{ST,PT,T2T},CLOperationRange{T}}},
    CLMixtureSpinSys{T,ST,PT,T2T}} where {T,ST,PT,T2T}

Takes in deserialized quantities to output a surrogate model and its corresponding mixture spin system data structure.